+++ spec/ubersicht/ingestion/client_spec.rb
@@ -9,6 +9,7 @@ RSpec.describe Ubersicht::Ingestion::Client do
 
   def build_options(custom = {})
     {
+      debug: true,
       hmac_key: hmac_key,
       account_id: account_id,
       pass: 'pass',
@@ -37,7 +38,7 @@ RSpec.describe Ubersicht::Ingestion::Client do
     it { expect(described_class.default(build_options)).to be_a(described_class) }
   end
 
-  describe '.ingest_events' do
+  describe '.ingest' do
     let(:event) { build_event }
     let(:client) { build_client }
 
@@ -64,7 +65,8 @@ RSpec.describe Ubersicht::Ingestion::Client do
       stub_request(:post, build_url)
         .to_return(status: Ubersicht::Ingestion::ACCEPTED_STATUS, body: Ubersicht::Ingestion::ACCEPTED_RESPONSE)
 
-      client.ingest(event.transaction_type, event.event_code, event.event_date, event.payload)
+      event_attrs = event.attributes.except(:payload).merge(event.payload)
+      expect(client.ingest(**event_attrs)).to eq(Ubersicht::Ingestion::ACCEPTED_RESPONSE)
 
       body = {
         events: [
@@ -86,21 +88,31 @@ RSpec.describe Ubersicht::Ingestion::Client do
       expect(WebMock).to have_requested(:post, build_url).with(body: body.to_json, headers: headers)
     end
 
-    it 'rejects invalid event' do
-      message = /:event_code violates constraints/
-      expect do
-        client.ingest(build_event[:transaction_type], nil, nil)
-      end.to raise_error(Ubersicht::ValidationError, message)
+    it 'ingests with debug=false' do
+      allow_any_instance_of(Ubersicht::Ingestion::HmacValidator).to receive(:calculate_notification_hmac)
+        .with(any_args, hmac_key)
+      stub_request(:post, build_url)
+        .to_return(status: Ubersicht::Ingestion::ACCEPTED_STATUS, body: Ubersicht::Ingestion::ACCEPTED_RESPONSE)
+
+      client = build_client(debug: false)
+      expect(client.ingest(**event.attributes)).to be_nil
+    end
+
+    it 'rejects event without event_code' do
+      event_attrs = build_event.attributes.except(:event_code)
+      expect { client.ingest(**event_attrs) }.to raise_error(ArgumentError, 'missing keyword: :event_code')
+    end
+
+    it 'rejects event without transaction_type' do
+      event_attrs = build_event.attributes.except(:transaction_type)
+      expect { client.ingest(**event_attrs) }.to raise_error(ArgumentError, 'missing keyword: :transaction_type')
     end
 
     it 'responds with exception' do
       stub_request(:post, build_url)
         .to_return(status: Ubersicht::Ingestion::REJECTED_STATUS, body: Ubersicht::Ingestion::REJECTED_RESPONSE)
 
-      expect do
-        client.ingest(event.transaction_type, event.event_code,
-                      event.event_date)
-      end.to raise_error(Ubersicht::Error, /rejected/)
+      expect { client.ingest(**event.attributes) }.to raise_error(Ubersicht::Error, /rejected/)
     end
   end
 end