+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 - [ingestion] added `event.payload.event_id` to prevent events duplication on Ubersicht side
 - [ingestion] added `event.payload.event_id` to event signature
+- [ingestion] added `debug` option to ingestion client. If false then request will be ran in a new thread.
 
 Braking changes:
 
-- [ingestion] moved `event.transaction_id` to `event.transaction.event_group_id`
+- [ingestion] replaced `ingest` signature to keywords
 - [ingestion] renamed `event.type` with `event.transaction_type`
 - [ingestion] replaced `clien.ingest_events(events)` with `clien.ingest(event)`